Updates 3
I turned the old references to different pages that would work with HTML into anchors here with ASP.NET and left the old references as comments. I've now only commenced to create a view for "Rings" but my goal here was to understand how to redirect a user to another page and that goal was met. Now it will be continuous work on the views for these categories. Then the sub categories. then the check out and so on. I also linked my database to this project as an update.