New DB platform, routing, and pages.
This is a more set platform for my database. I created the models based on it. I also implemented routing within in the controller for a nice URL view. Its simple an easy to read to the user. I created initial pages for all categories. Pretty much just a header for each. I have been using a Udemy course and the MSSA program for this and its been very helpful. As we do something in class I watch that portion of the Udemy course due to the fact that I am creating this in asp.net core and we are using mvc in class.